+++ .env.example
@@ -1,15 +1,15 @@
 
 # Environment Variables
-GEMINI_API_KEY=
+GEMINI_API_KEY=YOUR_GEMINI_KEY
 MODEL_NAME=gemini-2.5-flash
 MAX_TOKENS=1000
 TEMPERATURE=0.7
 
 # Firebase Configuration - skatit-ec470 Project
 FIREBASE_PROJECT_ID=skatit-ec470
-FIREBASE_PRIVATE_KEY_ID=9925ccbc23022c99dbd7b1f87008d64f5c245079
-FIREBASE_PRIVATE_KEY=
-FIREBASE_CLIENT_EMAIL=firebase-adminsdk-fbsvc@skatit-ec470.iam.gserviceaccount.com
-FIREBASE_CLIENT_ID=117936363127878834369
+FIREBASE_PRIVATE_KEY_ID=PRIVATE_KEY
+FIREBASE_CLIENT_EMAIL=CLIENT_EMAIL
+FIREBASE_CLIENT_ID=CLIENT_ID
+FIREBASE_CREDENTIALS_BASE64=BASE64_STRING_OF_FIREBASE_ADMINSDK_JSON
