+++ function/data.py
@@ -1,5 +1,5 @@
-from typing import List, Optional, Dict, Any
-from datetime import datetime, date
+from typing import List, Optional
+from datetime import datetime
 from pydantic import BaseModel, Field
