+++ function/managers/crisis.py
