+++ function/managers/events.py
