+++ function/managers/firebase_manager.py
